requirements.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
